+++ releasenotes.md
@@ -5,7 +5,7 @@
 ### Changes / improvements
 - vectors may now contain pointers.
 - `void!` does not convert to `anyfault`.
-- `$$masked_load` / `$$masked_store` for vector masked load/store. 
+- `$$masked_load` / `$$masked_store` / `$$gather` / `$$scatter` for vector masked load/store. 
 - `$$select` builtin for vector masked select.
 - Subtype matching in type switches.
 - Added parentof typeid property.